Dejo pendiente la elección de pokemones cuando uno de ellos es derrotado
